call-bq-with-python/requirements.txt
google-cloud-bigquery
matplotlib